SideBar component created/ added to homePage
