+++ src/components/index.js
@@ -1,2 +1,3 @@
 export * from "./Grid";
 export * from "./Cell";
+export * from "./SideBar";
